+++ CHANGES.rst
@@ -6,9 +6,18 @@ There's a frood who really knows where his towel is.
 1.6 (unreleased)
 ^^^^^^^^^^^^^^^^
 
-- Display results as votes or percentage (closes `#4`_). [marcosfromero]
+- Display results as votes or percentage (closes `#4`_).
+  [marcosfromero]
 
-- In portlet link back to poll page (closes `#47`_). [marcosfromero]
+- Add a default value to the voting portlet ``poll`` parameter. This fixes an
+  issue happening when importing ``portlets.xml`` assignments.
+  [ericof]
+
+- Remove unnecesary code in portlet (closes `#73`_).
+  [nueces]
+
+- In portlet link back to poll page (closes `#47`_).
+  [marcosfromero]
 
 
 1.5 (2013-12-17)
@@ -179,3 +188,4 @@ There's a frood who really knows where his towel is.
 .. _`#53`: https://github.com/collective/collective.polls/issues/53
 .. _`#61`: https://github.com/collective/collective.polls/issues/61
 .. _`#64`: https://github.com/collective/collective.polls/issues/64
+.. _`#73`: https://github.com/collective/collective.polls/issues/73